Override sqlite3 4.0.7 dependency with sqlite3 5.0.2
sqlite3 4.0.7 (a dependency of better-queue-sqlite) is a problematic
package, because a) there are no pre-compiled binaries for Node 14 and
b) it includes .vscode directory that takes up 151 Mb and clearly is
there by mistake. Using sqlite 5.0.2 results in a significantly faster
yarn install and docker build, smaller node_modules and
better-queue-sqlite appears to be working with it as well as it ever
did.