replace double bake/endorse checks with balance_updates based impl
Accusations metadata includes balance updates which specify
offending baker. Previous implementation was looking at baking rights
at accused level for double bake and looking for matching endorsement
signature in accused block operations for double endorsement. This
required extra rpc call. Note that metadata may not be included in the
accusation if we examine a block that is too old for node's history
mode, but the same goes for baking rights used in previous
implementation. Also note that analyzing old blocks is outside of
scope of this tool which is concerned with present and very recent
past.